# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "DeguBase – workspace and table management API.",
        "title": "DeguBase API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/workspaces/{wsCode}/scripts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scripts"
                ],
                "summary": "List scripts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace code",
                        "name": "wsCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Script"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scripts"
                ],
                "summary": "Create script",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace code",
                        "name": "wsCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "script",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/automation.CreateScriptRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Script"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/workspaces/{wsCode}/scripts/env": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scripts"
                ],
                "summary": "List script env vars",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace code",
                        "name": "wsCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ScriptEnvVar"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scripts"
                ],
                "summary": "Upsert script env var",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace code",
                        "name": "wsCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "env var",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/automation.UpsertEnvRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ScriptEnvVar"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/workspaces/{wsCode}/scripts/env/{envID}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "scripts"
                ],
                "summary": "Delete script env var",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace code",
                        "name": "wsCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "env var ID",
                        "name": "envID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/workspaces/{wsCode}/scripts/{scriptID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scripts"
                ],
                "summary": "Get script",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace code",
                        "name": "wsCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "script ID",
                        "name": "scriptID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Script"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scripts"
                ],
                "summary": "Update script",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace code",
                        "name": "wsCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "script ID",
                        "name": "scriptID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "script",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/automation.CreateScriptRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Script"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "scripts"
                ],
                "summary": "Delete script",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace code",
                        "name": "wsCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "script ID",
                        "name": "scriptID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/workspaces/{wsCode}/scripts/{scriptID}/executions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scripts"
                ],
                "summary": "List script executions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace code",
                        "name": "wsCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "script ID",
                        "name": "scriptID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "max results",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ScriptExecution"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scripts"
                ],
                "summary": "Record script execution",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace code",
                        "name": "wsCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "script ID",
                        "name": "scriptID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "execution result",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ScriptExecution"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.ScriptExecution"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "automation.CreateScriptRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "event_type": {
                    "$ref": "#/definitions/models.ScriptEventType"
                },
                "name": {
                    "type": "string"
                },
                "table_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "automation.UpsertEnvRequest": {
            "type": "object",
            "properties": {
                "is_secret": {
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "not_found"
                },
                "error": {
                    "type": "string",
                    "example": "not found"
                }
            }
        },
...
        "models.Script": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "event_type": {
                    "$ref": "#/definitions/models.ScriptEventType"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "table_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "workspace_id": {
                    "type": "integer"
                }
            }
        },
        "models.ScriptEnvVar": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "is_secret": {
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                },
                "workspace_id": {
                    "type": "integer"
                }
            }
        },
        "models.ScriptEventType": {
            "type": "string",
            "enum": [
                "record.created",
                "record.updated",
                "record.deleted"
            ],
            "x-enum-varnames": [
                "ScriptEventRecordCreated",
                "ScriptEventRecordUpdated",
                "ScriptEventRecordDeleted"
            ]
        },
        "models.ScriptExecution": {
            "type": "object",
            "properties": {
                "depth": {
                    "type": "integer"
                },
                "duration_ms": {
                    "type": "integer"
                },
                "ended_at": {
                    "type": "string"
                },
                "event_type": {
                    "$ref": "#/definitions/models.ScriptEventType"
                },
                "id": {
                    "type": "integer"
                },
                "logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ScriptLogEntry"
                    }
                },
                "row_id": {
                    "type": "integer"
                },
                "script_id": {
                    "type": "integer"
                },
                "started_at": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "table_code": {
                    "type": "string"
                },
                "triggered_by_execution_id": {
                    "type": "integer"
                }
            }
        },
        "models.ScriptLogEntry": {
            "type": "object",
            "properties": {
                "level": {
                    "$ref": "#/definitions/models.ScriptLogLevel"
                },
                "message": {
                    "type": "string"
                },
                "ts": {
                    "type": "string"
                }
            }
        },
        "models.ScriptLogLevel": {
            "type": "string",
            "enum": [
                "info",
                "warn",
                "error"
            ],
            "x-enum-varnames": [
                "ScriptLogInfo",
                "ScriptLogWarn",
                "ScriptLogError"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}